requirements.txt
anthropic
openai
beautifulsoup4
requests
playwright
pandas
openpyxl
python-dotenv
